Deprecate use expanded tile templates option
