+++ src/libs/components/expanded-tile-swiper/index.ts
@@ -21,11 +21,6 @@ export interface ExpandedTileSettings {
    */
   useDefaultAddToCartStyles: boolean
   /**
-   * Use default expanded tile templates
-   * @default true
-   */
-  useDefaultExpandedTileTemplates: boolean
-  /**
    * Default font - can be a google font link or an external font link
    * @default "https://fonts.googleapis.com/css2?family=Inter:wght@400;700&display=swap"
    */
@@ -42,6 +37,8 @@ export interface ExpandedTileSettings {
 }
 
 function loadDefaultExpandedTileTemplates(addExpandedTileTemplates: boolean) {
+  console.log("tile enabled", addExpandedTileTemplates)
+
   if (addExpandedTileTemplates) {
     sdk.addTemplateToComponent(ExpandedTiles, "expanded-tiles")
   }
@@ -55,11 +52,16 @@ function loadWidgetFonts(defaultFont: string) {
   }`)
 }
 
-export function loadExpandedTileTemplates(settings: ExpandedTileSettings) {
-  loadDefaultExpandedTileTemplates(settings.useDefaultExpandedTileTemplates)
+export function loadExpandedTileTemplates(settings: ExpandedTileSettings, templateEnabled: boolean) {
+  loadDefaultExpandedTileTemplates(templateEnabled)
   loadWidgetFonts(settings.defaultFont)
 
   if (settings.useDefaultSwiperStyles) {
     loadSwiperStyles()
   }
 }
+
+export * from "./tile.template"
+export * from "./embed-youtube.template"
+export * from "./video.templates"
+export * from "./types"
