+++ src/libs/components/expanded-tile-swiper/types.ts
@@ -1,0 +1,12 @@
+import { ISdk, Tile } from "../../../types"
+
+export type ExpandedTileProps = {
+  sdk: ISdk
+  tile: Tile
+}
+
+export type ShopspotProps = {
+  shopspotEnabled: boolean
+  parent?: string
+  tileId: string
+}
