+++ src/widget-loader.spec.tsx
@@ -33,7 +33,6 @@ const settings: EnforcedWidgetSettings<unknown> = {
       useDefaultExpandedTileStyles: true,
       useDefaultProductStyles: true,
       useDefaultAddToCartStyles: true,
-      useDefaultExpandedTileTemplates: true,
       defaultFont: "https://fonts.googleapis.com/css2?family=Inter:wght@400;700&display=swap",
       useDefaultSwiperStyles: true
     }
